added image for empty transaction list

+++ lib/models/transaction.dart
@@ -5,7 +5,7 @@ import 'package:flutter/foundation.dart';
 class Transaction {
   final String? id;
   final String? title;
-  final double? amount;
+  final int? amount;
 
   DateTime? date;
